+++ packages/core/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap/guided-answers-extension-core
 
+## 1.20.6
+
+### Patch Changes
+
+-   @sap/guided-answers-extension-types@1.20.6
+
+## 1.20.5
+
+### Patch Changes
+
+-   Updated dependencies [be0f952]
+    -   @sap/guided-answers-extension-types@1.20.5
+
 ## 1.20.4
 
 ### Patch Changes